Update README.md
##Project Summary
Airbnb is an American Company since 2007, it is an online marketplace that connects people who want to rent out their homes with people who are looking for accommodations in specific locales.

We have our dataset from AirBnb based on NY. NY is amongst the most expensive places to live in USA. We would like to perform an in-depth analysis on one of the most densely populated cities of world. Our dataset is feature rich containing, location with co-ordinates, prices, host name, room types, availability throughout season. From these features we’ve tried to extract information like most expensive places to live in NY, is location really varies with occupancy rate, what type of room people tends to choose most, is there any particular season for tourists or locale when we can follow a surge in prices or occupancy rate of properties etc.

We have performed data wrangling, handled n/a values, created new feature from last reviewed date. Also, we’ve plotted location wise data and distribution of our numerical features, performed univariate and bivariate analysis with multiple dependencies. Though we haven’t handled outlier in our analysis, also focused on neighborhood groups or large area of NY rather than small neighborhoods. We weren’t able to perform sentiment analysis or property quality analysis due to the lack of data, but can conclude that Manhattan and Brooklyn are of the most expensive areas of NY, people live a lavish life and tends not to prefer shared rooms even if shared rooms are the cheaper. We can understand in the middle of the year there is a surge in occupancy rate. Also, location has a high relation on deciding a property price but that doesn’t mean property in popular location will stay occupied on most of the season.